requirements.txt
absl-py==2.4.0
amqp==5.3.1
attrs==25.4.0
Automat==25.4.16
billiard==4.2.4
boto3==1.42.42
botocore==1.42.42
brotli==1.2.0
celery==5.6.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
contourpy==1.3.3
cryptography==46.0.4
cssselect==1.4.0
cycler==0.12.1
defusedxml==0.7.1
ffmpeg-python==0.2.0
filelock==3.20.3
Flask==3.0.0
Flask-Cors==4.0.0
flatbuffers==25.12.19
fonttools==4.61.1
fsspec==2026.2.0
future==1.0.0
greenlet==3.3.1
h5py==3.15.1
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
iniconfig==2.3.0
itemadapter==0.13.1
itemloaders==1.4.0
jax==0.9.0
jaxlib==0.9.0
Jinja2==3.1.6
jmespath==1.1.0
kiwisolver==1.4.9
kombu==5.6.2
loguru==0.7.3
lxml==6.0.2
MarkupSafe==3.0.3
matplotlib==3.10.8
mediapipe==0.10.32
ml_dtypes==0.5.4
mpmath==1.3.0
mutagen==1.47.0
networkx==3.6.1
numpy==2.4.2
opencv-contrib-python==4.13.0.92
opencv-python==4.13.0.92
opt_einsum==3.4.0
packaging==26.0
parsel==1.11.0
pillow==12.1.0
pluggy==1.6.0
polars==1.38.0
polars-runtime-32==1.38.0
prompt_toolkit==3.0.52
Protego==0.6.0
protobuf==4.25.8
psutil==7.2.2
psycopg2-binary==2.9.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pycryptodomex==3.23.0
PyDispatcher==2.0.7
Pygments==2.19.2
pyOpenSSL==25.3.0
pyparsing==3.3.2
PySide6==6.10.2
PySide6_Addons==6.10.2
PySide6_Essentials==6.10.2
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
queuelib==1.9.0
redis==5.0.1
requests==2.32.5
requests-file==3.0.1
s3transfer==0.16.0
scipy==1.17.0
Scrapy==2.14.1
service-identity==24.2.0
setuptools==80.10.2
shiboken6==6.10.2
six==1.17.0
sounddevice==0.5.5
SQLAlchemy==2.0.46
sympy==1.14.0
tldextract==5.3.1
torch==2.10.0
torchvision==0.25.0
Twisted==25.5.0
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
ultralytics==8.4.12
ultralytics-thop==2.0.18
urllib3==2.6.3
vine==5.1.0
w3lib==2.4.0
wcwidth==0.5.3
websockets==16.0
win32_setctime==1.2.0
yt-dlp==2026.2.4
zope.interface==8.2
